Configured app for localization. Updated audio player logic.
Added the basis for internationalizing the application. Starting with Spanish and English to learn about this workflow.

Updated the error reporting of the AVPlayer to check the underlying item's status when AVPlayerStatusFailed is reported. It's possible the player is OK but the underlying item is not.

Added method to get and print metadata for debugging.

Added NSLog to DLog macro.