Fixed Streaming Issues. Added Request & About Features.
Fixed issue with stream not playing by verifying the AVAudioPlayer and not the under lying AVAudioItem.

Added a UIActionSheet to allow the user to send requests to the station.

Set framework to receive and process stream metadata if/when the station provides it.

Added a 3rd view controller to display information about the station.